+++ CHANGES.rst
@@ -5,6 +5,8 @@
 - Use SYSDATE instead of NOW().
   Thanks `bouk <https://github.com/bouk>`_.
   (`Issue #15 <https://github.com/graingert/redshift_sqlalchemy/pull/15>`_)
+- Default to SSL with hardcoded AWS Redshift CA.
+  (`Issue #20 <https://github.com/graingert/redshift_sqlalchemy/pull/20>`_)
 
 
 0.1.2 (2015-08-11)