+++ CHANGES.rst
@@ -19,6 +19,8 @@
 - Add support for Redshift ``DELETE`` statements that refer other tables in the ``WHERE``
   clause. (`Issue #35 <https://github.com/graingert/redshift_sqlalchemy/issues/35>`_).
   Thanks `haleemur <https://github.com/haleemur> _`.
+- Raise ``NoSuchTableError`` when trying to reflect a table that doesn't exist.
+  (`Issue #38 <https://github.com/graingert/redshift_sqlalchemy/issues/38>`_).
 
 0.1.2 (2015-08-11)
 ------------------