+++ CHANGES.rst
@@ -1,5 +1,5 @@
 
-0.1.3 (unreleased)
+0.2.0 (2015-09-04)
 ------------------
 
 - Use SYSDATE instead of NOW().